white_belt/requirements.txt
fastapi>=0.68.0,<0.69.0
#pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
google-cloud-storage==1.42.3
#geopandas==0.6.1
netCDF4
shapely
pandas
google-generativeai
python-dotenv